backend_fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
mysqlclient
PyJWT
pymysql